Created CSS for Strain Details
Had to copy over the HTML of your Strain Detail page and make some
slight adjustments so that I can view what I was doing -- suppose you'll
need to re-route where the image and stylesheet are pulled from -- but I
created a stylesheet that more or less pushes the content to resemble
pg. 3 of your wireframe.