About Us Placeholder
Created some placeholder content for the About Us page -- probably more
convoluted than we'd ultimately like it to be but I know we need this
done tonight so perhaps we can smooth it out later.

As with the Strain Detail, the CSS and images will need to be sourced
differently when you implement them. I created a separate stylesheet for
the About Us page until I can take time to incorporate it into the
stylesheet for the overall site.